Add a function to check if job has completed

+++ start.py
@@ -83,6 +83,23 @@ cmd_out = cmd.stdout.read()
 cmd_err = cmd.stderr.read()
 print(cmd_out.decode())
 print(cmd_err.decode())
+
+
+def check_squeue():
+    cmd = subprocess.Popen("gcloud compute ssh g1-login1 --zone="+zone+" --command 'squeue'",
+                           shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    cmd.wait()
+    cmd_out = cmd.stdout.read()
+    cmd_err = cmd.stderr.read()
+    print(cmd_out.decode())
+    print(cmd_err.decode())
+
+
+if cmd_out.decode() != "             JOBID PARTITION     NAME     USER ST       TIME  NODES NODELIST(REASON)":
+    print("Waiting for job to finish")
+    time.sleep(10)
+    check_squeue()
+
 cmd = subprocess.Popen("gcloud compute scp g1-login1:~/code/test-srun.out . --zone=" +
                        zone, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 cmd.wait()
@@ -90,12 +107,21 @@ cmd_out = cmd.stdout.read()
 cmd_err = cmd.stderr.read()
 print(cmd_out.decode())
 print(cmd_err.decode())
-quit()
-print("Destroying Slurm Cluster")
-cmd = subprocess.Popen("gcloud deployment-manager deployments delete google1 -q",
-                       shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-cmd.wait()
-cmd_out = cmd.stdout.read()
-cmd_err = cmd.stderr.read()
-print(cmd_out.decode())
-print(cmd_err.decode())
+
+answer = None
+while answer not in ("yes", "no"):
+    answer = input(
+        "Job executed successfully. Would you like to delete the deployment?\n Enter yes or no:")
+    if answer == "yes":
+        quit()
+    elif answer == "no":
+        print("Destroying Slurm Cluster")
+        cmd = subprocess.Popen("gcloud deployment-manager deployments delete google1 -q",
+                               shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        cmd.wait()
+        cmd_out = cmd.stdout.read()
+        cmd_err = cmd.stderr.read()
+        print(cmd_out.decode())
+        print(cmd_err.decode())
+    else:
+        print("Please enter yes or no.")
